[fix] Bcrypt 를 이용해 비밀번호 암호화
1. 기존 평문으로 저장하던 비밀번호를 Bcrypt를 이용해 암호화함

+++ member-api/member-application/src/main/java/org/flab/deliveryplatform/member/application/service/WithdrawMemberService.java
@@ -5,6 +5,7 @@ import org.flab.deliveryplatform.member.application.port.MemberRepository;
 import org.flab.deliveryplatform.member.application.port.WithdrawMemberUseCase;
 import org.flab.deliveryplatform.member.application.port.dto.WithdrawMemberCommand;
 import org.flab.deliveryplatform.member.application.port.exception.InvalidMemberInfoException;
+import org.flab.deliveryplatform.member.application.service.utils.EncryptUtils;
 import org.flab.deliveryplatform.member.domain.Member;
 import org.springframework.stereotype.Service;
 
@@ -14,11 +15,16 @@ public class WithdrawMemberService implements WithdrawMemberUseCase {
 
     private final MemberRepository memberRepository;
 
+    private final EncryptUtils encryptUtils;
+
     @Override
     public void withdraw(WithdrawMemberCommand withdrawMemberCommand) {
-        Member member = memberRepository.findByEmailAndPassword(
-                withdrawMemberCommand.getEmail(), withdrawMemberCommand.getPassword())
+        Member member = memberRepository.findByEmail(withdrawMemberCommand.getEmail())
             .orElseThrow(InvalidMemberInfoException::new);
+
+        if (!encryptUtils.isMatch(withdrawMemberCommand.getPassword(), member.getPassword())) {
+            throw new InvalidMemberInfoException();
+        }
 
         memberRepository.delete(member);
     }
